mock_service/requirements-dev.txt
grpcio>=1.44.0
protobuf>=3.19.4
types-protobuf
grpcio-tools
pytest
